+++ src/models/bundle.rs
@@ -1,0 +1,20 @@
+use serde::Deserialize;
+use tabled::Tabled;
+
+use crate::models::null_to_default;
+
+#[allow(dead_code)]
+#[derive(Debug, Tabled, Deserialize)]
+pub struct Bundles {
+    pub built_in: bool,
+    #[serde(deserialize_with = "null_to_default")]
+    pub created_time: i64,
+    #[serde(deserialize_with = "null_to_default")]
+    pub description: String,
+    #[serde(deserialize_with = "null_to_default")]
+    pub id: i64,
+    #[serde(deserialize_with = "null_to_default")]
+    pub mod_time: i64,
+    #[serde(deserialize_with = "null_to_default")]
+    pub name: String,
+}
