+++ src/models/mod.rs
@@ -1,0 +1,12 @@
+use serde::{Deserialize, Deserializer};
+
+pub mod bundle;
+
+pub fn null_to_default<'de, D, T>(deserializer: D) -> Result<T, D::Error>
+where
+    T: Default + Deserialize<'de>,
+    D: Deserializer<'de>,
+{
+    let opt = Option::deserialize(deserializer)?;
+    Ok(opt.unwrap_or_default())
+}
